Optimize code preparing to merge into master branch
