+++ .gitignore
@@ -10,3 +10,4 @@ compile_flags.txt
 doc
 mined
 scancode_error.txt
+.vscode
