+++ src/url.c
@@ -34,9 +34,8 @@
 #include <stdlib.h>
 #include <string.h>
 #include "file.h"
-//#include "minr.h"
+#include "minr.h"
 #include "ldb.h"
-#include "scancode.h"
 
 /**
  * @brief Calculate purl md5
@@ -180,8 +179,8 @@ void url_download(struct minr_job *job)
 		/* Add info to urls.csv */
 		url_add(job);
 		
-		if (job->scancode_mode)
-			scancode_mine_attribution_notice(job);
+		/* Find for license declaration into specific files in the roor dir (no recursive) */
+		mine_license_exec(job);
 
 		recurse(job, job->tmp_dir);
